Apply Higher-Order Components
